requirements.txt
celery==5.3.6
google-auth==2.29.0
google-api-python-client==2.125.0
python-dotenv==1.0.1